+++ include/cifsdtools.h
@@ -108,7 +108,7 @@ extern void pr_logger_init(int flags);
 
 void pr_hex_dump(const void *mem, size_t sz);
 
-char *base64_encode(unsigned char *src, unsigned long srclen);
-unsigned char *base64_decode(char const *src, unsigned long *dstlen);
+char *base64_encode(unsigned char *src, size_t srclen);
+unsigned char *base64_decode(char const *src, size_t *dstlen);
 
 #endif /* __CIFSDTOOLS_H__ */
